fix(dashboard): preserve lazy recovery bundle
