+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
   "maximum_collected_cases": 15541,
-  "maximum_test_source_lines": 335037,
+  "maximum_test_source_lines": 335039,
   "schema_version": 1
 }
