+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/app-detail-workspace.js
@@ -1,4 +1,4 @@
-import { r as reactExports, j as jsxRuntimeExports, aP as fetchApprovalPage, aQ as fetchPolicy, v as guardActionDisposition, p as protectionHealthFor, k as useProtectionPresentationState, aL as HiMiniArrowLeft, c as HiMiniChevronRight, i as harnessDisplayName, q as GuardHero, R as ProofStrip, aR as HiMiniHome, N as HiMiniBolt, a6 as HiMiniAdjustmentsHorizontal, S as SectionLabel, A as ActionButton, t as HiMiniShieldCheck, w as formatRelativeTime, M as HiMiniExclamationTriangle, aS as guardActionPresentation, P as Badge, aT as DEFAULT_FILTER_STATE, aU as filterEvidence, aV as sortEvidence, aW as computeMetrics, aX as CommandActivityWorkspace, m as EmptyState, aY as EvidenceFilterBar, aZ as EvidenceInsightStrip, a_ as EvidenceActionList, a$ as EvidenceActionDetail, K as useFocusTrap, b0 as policyIdentityKey, I as HiMiniCloud, b1 as HiMiniChartBar, at as Tag, o as HiMiniCheckCircle, X as HiMiniXCircle, b2 as runHarnessAction, b3 as GuardHarnessActionError, b4 as HiMiniRocketLaunch, ay as HiMiniArrowPath, b5 as HiMiniTrash, b6 as clearLabelForScope, b7 as formatHarnessCommand } from "../guard-dashboard.js";
+import { r as reactExports, j as jsxRuntimeExports, aS as fetchApprovalPage, aT as fetchPolicy, v as guardActionDisposition, p as protectionHealthFor, k as useProtectionPresentationState, aO as HiMiniArrowLeft, c as HiMiniChevronRight, i as harnessDisplayName, q as GuardHero, Z as ProofStrip, aU as HiMiniHome, N as HiMiniBolt, a9 as HiMiniAdjustmentsHorizontal, S as SectionLabel, A as ActionButton, t as HiMiniShieldCheck, w as formatRelativeTime, M as HiMiniExclamationTriangle, aV as guardActionPresentation, P as Badge, aW as DEFAULT_FILTER_STATE, aX as filterEvidence, aY as sortEvidence, aZ as computeMetrics, a_ as CommandActivityWorkspace, m as EmptyState, a$ as EvidenceFilterBar, b0 as EvidenceInsightStrip, b1 as EvidenceActionList, b2 as EvidenceActionDetail, K as useFocusTrap, b3 as policyIdentityKey, I as HiMiniCloud, b4 as HiMiniChartBar, aw as Tag, o as HiMiniCheckCircle, $ as HiMiniXCircle, b5 as runHarnessAction, b6 as GuardHarnessActionError, b7 as HiMiniRocketLaunch, aB as HiMiniArrowPath, b8 as HiMiniTrash, b9 as clearLabelForScope, ba as formatHarnessCommand } from "../guard-dashboard.js";
 import { a as appSetupTarget } from "./harness-setup-target.js";
 function ActivityModeButton(props) {
   const active = props.mode === props.value;
