+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/approval-proof-modal.js
@@ -1,4 +1,4 @@
-import { r as reactExports, aC as buildApprovalProofCredentials, aA as isApprovalProofSubmitDisabled, j as jsxRuntimeExports, S as SectionLabel, aB as ApprovalProofFieldInputs, A as ActionButton } from "../guard-dashboard.js";
+import { r as reactExports, aF as buildApprovalProofCredentials, aD as isApprovalProofSubmitDisabled, j as jsxRuntimeExports, S as SectionLabel, aE as ApprovalProofFieldInputs, A as ActionButton } from "../guard-dashboard.js";
 function ApprovalProofModal(props) {
   const { title, detail, confirmLabel, approvalGate, busy = false, error = null, onCancel, onConfirm } = props;
   const [password, setPassword] = reactExports.useState("");
