+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/audit-workspace.js
@@ -1,4 +1,4 @@
-import { j as jsxRuntimeExports, at as Tag, w as formatRelativeTime, r as reactExports, A as ActionButton, c4 as HiMiniChevronLeft, c as HiMiniChevronRight, bd as IconActionButton, z as HiMiniXMark, bi as GuardModalLayer, c5 as HiMiniFunnel, aq as HiMiniMagnifyingGlass, c6 as HiMiniArrowDown, c7 as HiMiniArrowUp, S as SectionLabel, ay as HiMiniArrowPath, bh as HiMiniBugAnt, m as EmptyState, a6 as HiMiniAdjustmentsHorizontal, bj as ConnectFlowCard, M as HiMiniExclamationTriangle, ax as useResolvedApprovalGate, c8 as runAuditRemediation, P as Badge, c1 as isBlockedGuardAction, b9 as isSupplyChainAuditEvidence, o as HiMiniCheckCircle, X as HiMiniXCircle, i as harnessDisplayName, bD as HiMiniDocumentText, aO as guardAwareHref, t as HiMiniShieldCheck } from "../guard-dashboard.js";
+import { j as jsxRuntimeExports, aw as Tag, w as formatRelativeTime, r as reactExports, A as ActionButton, c6 as HiMiniChevronLeft, c as HiMiniChevronRight, bg as IconActionButton, z as HiMiniXMark, bl as GuardModalLayer, c7 as HiMiniFunnel, at as HiMiniMagnifyingGlass, c8 as HiMiniArrowDown, c9 as HiMiniArrowUp, S as SectionLabel, aB as HiMiniArrowPath, bk as HiMiniBugAnt, m as EmptyState, a9 as HiMiniAdjustmentsHorizontal, bm as ConnectFlowCard, M as HiMiniExclamationTriangle, aA as useResolvedApprovalGate, ca as runAuditRemediation, P as Badge, c3 as isBlockedGuardAction, bc as isSupplyChainAuditEvidence, o as HiMiniCheckCircle, $ as HiMiniXCircle, i as harnessDisplayName, bF as HiMiniDocumentText, aR as guardAwareHref, t as HiMiniShieldCheck } from "../guard-dashboard.js";
 import { A as ApprovalProofModal } from "./approval-proof-modal.js";
 import { p as packageWorkbenchEcosystems, f as filterPackageWorkbenchFindings, b as sortPackageWorkbenchFindings, i as isApprovalGateRequiredError } from "./supply-chain-hub-workspace.js";
 import { r as resolveManagerCoverageManagers, a as resolveManagerCoverageStatus } from "./supply-chain-protection-stats.js";
