+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/extensions-workspace.js
@@ -1,4 +1,4 @@
-import { av as fetchLocalCliApi, r as reactExports, aw as fetchExtensionControlApi, j as jsxRuntimeExports, ax as useResolvedApprovalGate, a4 as HiMiniLockClosed, M as HiMiniExclamationTriangle, ay as HiMiniArrowPath, t as HiMiniShieldCheck, az as HiMiniInformationCircle, aA as isApprovalProofSubmitDisabled, z as HiMiniXMark, aB as ApprovalProofFieldInputs, aC as buildApprovalProofCredentials, aD as GenIcon, N as HiMiniBolt, aE as HiMiniGlobeAlt, aF as HiMiniCube, I as HiMiniCloud, aG as HiMiniServerStack, b as HiMiniCommandLine, aH as HiMiniFolder, aI as FaWindows, aJ as FaAws, o as HiMiniCheckCircle, c as HiMiniChevronRight, C as HiMiniChevronDown, aK as approvalProofRecentlySatisfied, aL as HiMiniArrowLeft, aM as HiMiniPlus, Z as HiMiniClipboardDocumentCheck, _ as HiMiniClipboard, aq as HiMiniMagnifyingGlass, y as HiMiniSparkles, aN as HiMiniNoSymbol, ap as WorkspacePageHeader, aO as guardAwareHref } from "../guard-dashboard.js";
+import { ay as fetchLocalCliApi, r as reactExports, az as fetchExtensionControlApi, j as jsxRuntimeExports, aA as useResolvedApprovalGate, a7 as HiMiniLockClosed, M as HiMiniExclamationTriangle, aB as HiMiniArrowPath, t as HiMiniShieldCheck, aC as HiMiniInformationCircle, aD as isApprovalProofSubmitDisabled, z as HiMiniXMark, aE as ApprovalProofFieldInputs, aF as buildApprovalProofCredentials, aG as GenIcon, N as HiMiniBolt, aH as HiMiniGlobeAlt, aI as HiMiniCube, I as HiMiniCloud, aJ as HiMiniServerStack, b as HiMiniCommandLine, aK as HiMiniFolder, aL as FaWindows, aM as FaAws, o as HiMiniCheckCircle, c as HiMiniChevronRight, C as HiMiniChevronDown, aN as approvalProofRecentlySatisfied, aO as HiMiniArrowLeft, aP as HiMiniPlus, a0 as HiMiniClipboardDocumentCheck, a1 as HiMiniClipboard, at as HiMiniMagnifyingGlass, y as HiMiniSparkles, aQ as HiMiniNoSymbol, as as WorkspacePageHeader, aR as guardAwareHref } from "../guard-dashboard.js";
 import { A as ApprovalProofModal } from "./approval-proof-modal.js";
 const EXTENSION_ID_PATTERN = /^command\.[a-z0-9]+(?:[.-][a-z0-9]+)*$/;
 const RULE_ID_PATTERN = /^command\.[a-z0-9]+(?:[.-][a-z0-9]+)*$/;
