+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/feed-health-workspace.js
@@ -1,4 +1,4 @@
-import { j as jsxRuntimeExports, A as ActionButton, S as SectionLabel, M as HiMiniExclamationTriangle, ay as HiMiniArrowPath, o as HiMiniCheckCircle, c9 as HiMiniSignal, X as HiMiniXCircle, bc as HiMiniClock, w as formatRelativeTime, P as Badge, at as Tag } from "../guard-dashboard.js";
+import { j as jsxRuntimeExports, A as ActionButton, S as SectionLabel, M as HiMiniExclamationTriangle, aB as HiMiniArrowPath, o as HiMiniCheckCircle, cb as HiMiniSignal, $ as HiMiniXCircle, bf as HiMiniClock, w as formatRelativeTime, P as Badge, aw as Tag } from "../guard-dashboard.js";
 function resolveFeedSourceMode(cloudState) {
   if (cloudState === "local_only") return "sample";
   if (cloudState === "paired_waiting") return "full";
