+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/mcp-policy-request-panel.js
@@ -1,4 +1,4 @@
-import { r as reactExports, bX as fetchMcpPolicyRequest, aC as buildApprovalProofCredentials, bY as resolveMcpPolicyRequest, j as jsxRuntimeExports, m as EmptyState, A as ActionButton, ay as HiMiniArrowPath, aA as isApprovalProofSubmitDisabled, ap as WorkspacePageHeader, P as Badge, t as HiMiniShieldCheck, o as HiMiniCheckCircle, M as HiMiniExclamationTriangle, S as SectionLabel, bc as HiMiniClock, bZ as HiMiniDocumentPlus, b_ as HiMiniDocumentMagnifyingGlass, aN as HiMiniNoSymbol, aB as ApprovalProofFieldInputs, a3 as HiMiniKey } from "../guard-dashboard.js";
+import { r as reactExports, bZ as fetchMcpPolicyRequest, aF as buildApprovalProofCredentials, b_ as resolveMcpPolicyRequest, j as jsxRuntimeExports, m as EmptyState, A as ActionButton, aB as HiMiniArrowPath, aD as isApprovalProofSubmitDisabled, as as WorkspacePageHeader, P as Badge, t as HiMiniShieldCheck, o as HiMiniCheckCircle, M as HiMiniExclamationTriangle, S as SectionLabel, bf as HiMiniClock, b$ as HiMiniDocumentPlus, c0 as HiMiniDocumentMagnifyingGlass, aQ as HiMiniNoSymbol, aE as ApprovalProofFieldInputs, a6 as HiMiniKey } from "../guard-dashboard.js";
 const STATUS_LABELS = {
   pending: "Pending review",
   applied: "Applied",
