+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/policy-workspace-page.js
@@ -1,4 +1,4 @@
-import { j as jsxRuntimeExports, S as SectionLabel, z as HiMiniXMark, P as Badge, at as Tag, b as HiMiniCommandLine, M as HiMiniExclamationTriangle, bC as scopeLabel, i as harnessDisplayName, A as ActionButton, aO as guardAwareHref, w as formatRelativeTime$1, bD as HiMiniDocumentText, o as HiMiniCheckCircle, bE as HiMiniCloudArrowUp, bF as HiMiniCheck, bG as HiMiniCodeBracket, bH as HiMiniClipboardDocument, bI as HiMiniUsers, be as HiMiniBeaker, aH as HiMiniFolder, a4 as HiMiniLockClosed, t as HiMiniShieldCheck, az as HiMiniInformationCircle, bm as HiMiniCloudArrowDown, bl as HiMiniArrowTopRightOnSquare, bJ as HiMiniIdentification, bK as policyActionLabel, r as reactExports, bL as createCloudExceptionRequest, bM as HiMiniArrowRight, m as EmptyState, aq as HiMiniMagnifyingGlass, B as HiMiniChevronUp, C as HiMiniChevronDown, c as HiMiniChevronRight, bN as HiMiniPuzzlePiece, aE as HiMiniGlobeAlt, bc as HiMiniClock, bO as fetchCloudExceptions, bP as fetchCloudExceptionRequests, bQ as downloadBlob, bR as PolicyStatField, bS as PaginationControls, aN as HiMiniNoSymbol, aF as HiMiniCube, ay as HiMiniArrowPath, I as HiMiniCloud, a6 as HiMiniAdjustmentsHorizontal, bT as HiMiniArrowDownTray, bU as HiMiniQueueList, ap as WorkspacePageHeader, bA as lazyWorkspace, bB as __vitePreload } from "../guard-dashboard.js";
+import { j as jsxRuntimeExports, S as SectionLabel, z as HiMiniXMark, P as Badge, aw as Tag, b as HiMiniCommandLine, M as HiMiniExclamationTriangle, bE as scopeLabel, i as harnessDisplayName, A as ActionButton, aR as guardAwareHref, w as formatRelativeTime$1, bF as HiMiniDocumentText, o as HiMiniCheckCircle, bG as HiMiniCloudArrowUp, bH as HiMiniCheck, bI as HiMiniCodeBracket, bJ as HiMiniClipboardDocument, bK as HiMiniUsers, bh as HiMiniBeaker, aK as HiMiniFolder, a7 as HiMiniLockClosed, t as HiMiniShieldCheck, aC as HiMiniInformationCircle, bp as HiMiniCloudArrowDown, bo as HiMiniArrowTopRightOnSquare, bL as HiMiniIdentification, bM as policyActionLabel, r as reactExports, bN as createCloudExceptionRequest, bO as HiMiniArrowRight, m as EmptyState, at as HiMiniMagnifyingGlass, B as HiMiniChevronUp, C as HiMiniChevronDown, c as HiMiniChevronRight, bP as HiMiniPuzzlePiece, aH as HiMiniGlobeAlt, bf as HiMiniClock, bQ as fetchCloudExceptions, bR as fetchCloudExceptionRequests, bS as downloadBlob, bT as PolicyStatField, bU as PaginationControls, aQ as HiMiniNoSymbol, aI as HiMiniCube, aB as HiMiniArrowPath, I as HiMiniCloud, a9 as HiMiniAdjustmentsHorizontal, bV as HiMiniArrowDownTray, bW as HiMiniQueueList, as as WorkspacePageHeader, bC as lazyWorkspace, bD as __vitePreload } from "../guard-dashboard.js";
 const CLOUD_EXCEPTION_EXPIRING_SOON_DAYS = 7;
 function parseCloudExceptionTimestamp(value) {
   if (!value || !value.trim()) {
