+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/settings-workspace.js
@@ -1,4 +1,4 @@
-import { $ as PROTECTION_POSTURE_COPY, a0 as POSTURE_OUTCOME_COLUMNS, j as jsxRuntimeExports, r as reactExports, a1 as getDefaultExportFromCjs, a2 as React, K as useFocusTrap, a3 as HiMiniKey, S as SectionLabel, A as ActionButton, t as HiMiniShieldCheck, a4 as HiMiniLockClosed, a5 as HiMiniBellAlert, a6 as HiMiniAdjustmentsHorizontal, a7 as HiMiniCircleStack, a8 as TabBar, c as HiMiniChevronRight, a9 as resolveProtectionLevelCopy, aa as fetchSettings, ab as fetchRuntimeSnapshot, e as updateSettings, ac as clearPolicy, ad as clearReviewQueue, ae as revokeApprovalGateCooldown, af as disableApprovalGateTotp, ag as importSettings, ah as resetSettings, ai as enrollApprovalGateTotp, aj as verifyApprovalGateTotp, ak as clearEvidence, al as exportDiagnostics, am as repairApprovalCenter, an as exportSettings, ao as setupDesktopNotifications, m as EmptyState, ap as WorkspacePageHeader, W as WatchProtectionBanner, aq as HiMiniMagnifyingGlass, C as HiMiniChevronDown, o as HiMiniCheckCircle, M as HiMiniExclamationTriangle, ar as isProtectionPosture, as as deriveProtectionPosture, at as Tag, au as approvalGateCooldownLabel, z as HiMiniXMark } from "../guard-dashboard.js";
+import { a2 as PROTECTION_POSTURE_COPY, a3 as POSTURE_OUTCOME_COLUMNS, j as jsxRuntimeExports, r as reactExports, a4 as getDefaultExportFromCjs, a5 as React, K as useFocusTrap, a6 as HiMiniKey, S as SectionLabel, A as ActionButton, t as HiMiniShieldCheck, a7 as HiMiniLockClosed, a8 as HiMiniBellAlert, a9 as HiMiniAdjustmentsHorizontal, aa as HiMiniCircleStack, ab as TabBar, c as HiMiniChevronRight, ac as resolveProtectionLevelCopy, ad as fetchSettings, ae as fetchRuntimeSnapshot, e as updateSettings, af as clearPolicy, ag as clearReviewQueue, ah as revokeApprovalGateCooldown, ai as disableApprovalGateTotp, aj as importSettings, ak as resetSettings, al as enrollApprovalGateTotp, am as verifyApprovalGateTotp, an as clearEvidence, ao as exportDiagnostics, ap as repairApprovalCenter, aq as exportSettings, ar as setupDesktopNotifications, m as EmptyState, as as WorkspacePageHeader, W as WatchProtectionBanner, at as HiMiniMagnifyingGlass, C as HiMiniChevronDown, o as HiMiniCheckCircle, M as HiMiniExclamationTriangle, au as isProtectionPosture, av as deriveProtectionPosture, aw as Tag, ax as approvalGateCooldownLabel, z as HiMiniXMark } from "../guard-dashboard.js";
 import { f as filterSettingsBySearch, R as RISK_CONTROL_CONSEQUENCES } from "./app-catalog.js";
 const POSTURE_ORDER = ["protected", "extra_careful", "watch"];
 function ProtectionPosturePanel(props) {
