+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/supply-chain-hub-workspace.js
@@ -1,5 +1,5 @@
 const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/chunks/supply-chain-workspace.js","assets/guard-dashboard.js","assets/index.css","assets/chunks/feed-health-workspace.js","assets/chunks/home-protection-module.js","assets/chunks/supply-chain-protection-stats.js","assets/chunks/approval-proof-modal.js","assets/chunks/audit-workspace.js"])))=>i.map(i=>d[i]);
-import { b8 as isSupplyChainAuditIncomplete, b9 as isSupplyChainAuditEvidence, b3 as GuardHarnessActionError, ba as readString$1, bb as isRecord$1, r as reactExports, j as jsxRuntimeExports, o as HiMiniCheckCircle, ay as HiMiniArrowPath, M as HiMiniExclamationTriangle, at as Tag, w as formatRelativeTime, bc as HiMiniClock, bd as IconActionButton, X as HiMiniXCircle, b5 as HiMiniTrash, t as HiMiniShieldCheck, V as HiMiniWrenchScrewdriver, be as HiMiniBeaker, bf as ActivationSummary, bg as ActionResultPanel, aq as HiMiniMagnifyingGlass, m as EmptyState, A as ActionButton, bh as HiMiniBugAnt, z as HiMiniXMark, aC as buildApprovalProofCredentials, bi as GuardModalLayer, bj as ConnectFlowCard, bk as ApprovalProofInline, bl as HiMiniArrowTopRightOnSquare, bm as HiMiniCloudArrowDown, ax as useResolvedApprovalGate, bn as fetchPackageFirewallStatus, bo as runPackageAudit, bp as resolveSupplyChainAuditFailure, bq as runPackageSync, br as startPackageFirewallConnect, bs as openPackageFirewallAuthorizeFallback, bt as PACKAGE_FIREWALL_CONNECT_POPUP_BLOCKED_MESSAGE, bu as repairSupplyChainProtection, bv as runPackageFirewallAction, bw as parseInterceptProofSnapshot, bx as activatePackageFirewallRuntime, S as SectionLabel, by as EntitlementNotice, bz as fetchReceipts, ap as WorkspacePageHeader, bA as lazyWorkspace, bB as __vitePreload } from "../guard-dashboard.js";
+import { bb as isSupplyChainAuditIncomplete, bc as isSupplyChainAuditEvidence, b6 as GuardHarnessActionError, bd as readString$1, be as isRecord$1, r as reactExports, j as jsxRuntimeExports, o as HiMiniCheckCircle, aB as HiMiniArrowPath, M as HiMiniExclamationTriangle, aw as Tag, w as formatRelativeTime, bf as HiMiniClock, bg as IconActionButton, $ as HiMiniXCircle, b8 as HiMiniTrash, t as HiMiniShieldCheck, X as HiMiniWrenchScrewdriver, bh as HiMiniBeaker, bi as ActivationSummary, bj as ActionResultPanel, at as HiMiniMagnifyingGlass, m as EmptyState, A as ActionButton, bk as HiMiniBugAnt, z as HiMiniXMark, aF as buildApprovalProofCredentials, bl as GuardModalLayer, bm as ConnectFlowCard, bn as ApprovalProofInline, bo as HiMiniArrowTopRightOnSquare, bp as HiMiniCloudArrowDown, aA as useResolvedApprovalGate, bq as fetchPackageFirewallStatus, br as runPackageAudit, bs as resolveSupplyChainAuditFailure, bt as runPackageSync, bu as startPackageFirewallConnect, V as openPackageFirewallAuthorizeFallback, bv as PACKAGE_FIREWALL_CONNECT_POPUP_BLOCKED_MESSAGE, bw as repairSupplyChainProtection, bx as runPackageFirewallAction, by as parseInterceptProofSnapshot, bz as activatePackageFirewallRuntime, S as SectionLabel, bA as EntitlementNotice, bB as fetchReceipts, as as WorkspacePageHeader, bC as lazyWorkspace, bD as __vitePreload } from "../guard-dashboard.js";
 import { A as ApprovalProofModal } from "./approval-proof-modal.js";
 const SEVERITY_RANK = {
   critical: 4,
