+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/supply-chain-workspace.js
@@ -1,4 +1,4 @@
-import { r as reactExports, b$ as fetchSupplyChainBundle, j as jsxRuntimeExports, S as SectionLabel, m as EmptyState, bl as HiMiniArrowTopRightOnSquare, at as Tag, w as formatRelativeTime, P as Badge, M as HiMiniExclamationTriangle, bh as HiMiniBugAnt, aS as guardActionPresentation, c0 as isSupplyChainScannerEvidence, c1 as isBlockedGuardAction, ay as HiMiniArrowPath, b_ as HiMiniDocumentMagnifyingGlass, c2 as HiMiniShieldExclamation, c3 as HiMiniComputerDesktop, I as HiMiniCloud, o as HiMiniCheckCircle, V as HiMiniWrenchScrewdriver, A as ActionButton, C as HiMiniChevronDown, Y as HiMiniExclamationCircle, bz as fetchReceipts, X as HiMiniXCircle, i as harnessDisplayName, B as HiMiniChevronUp } from "../guard-dashboard.js";
+import { r as reactExports, c1 as fetchSupplyChainBundle, j as jsxRuntimeExports, S as SectionLabel, m as EmptyState, bo as HiMiniArrowTopRightOnSquare, aw as Tag, w as formatRelativeTime, P as Badge, M as HiMiniExclamationTriangle, bk as HiMiniBugAnt, aV as guardActionPresentation, c2 as isSupplyChainScannerEvidence, c3 as isBlockedGuardAction, aB as HiMiniArrowPath, c0 as HiMiniDocumentMagnifyingGlass, c4 as HiMiniShieldExclamation, c5 as HiMiniComputerDesktop, I as HiMiniCloud, o as HiMiniCheckCircle, X as HiMiniWrenchScrewdriver, A as ActionButton, C as HiMiniChevronDown, Y as HiMiniExclamationCircle, bB as fetchReceipts, $ as HiMiniXCircle, i as harnessDisplayName, B as HiMiniChevronUp } from "../guard-dashboard.js";
 import { resolveFeedStaleness } from "./feed-health-workspace.js";
 import { r as resolveHomeProtectionStatus } from "./home-protection-module.js";
 import { b as buildSupplyChainStats, r as resolveManagerCoverageManagers, a as resolveManagerCoverageStatus } from "./supply-chain-protection-stats.js";
